Started ptr-based pool impl
